solution3

+++ src/database/models/movies.py
@@ -1,7 +1,7 @@
 import datetime
 from enum import Enum
 from typing import Optional
-
+from decimal import Decimal
 from sqlalchemy import (
     String,
     Float,
@@ -131,8 +131,8 @@ class MovieModel(Base):
     status: Mapped[MovieStatusEnum] = mapped_column(
         SQLAlchemyEnum(MovieStatusEnum), nullable=False
     )
-    budget: Mapped[float] = mapped_column(DECIMAL(15, 2), nullable=False)
-    revenue: Mapped[float] = mapped_column(Float, nullable=False)
+    budget: Mapped[Decimal] = mapped_column(DECIMAL(15, 2), nullable=False)
+    revenue: Mapped[Decimal] = mapped_column(DECIMAL(15, 2), nullable=False)
 
     country_id: Mapped[int] = mapped_column(ForeignKey("countries.id"), nullable=False)
     country: Mapped["CountryModel"] = relationship("CountryModel", back_populates="movies")
